onboarding_project/requirements.txt
Django==4.2.16
pdfplumber==0.11.0
django-cors-headers==4.3.1
channels==4.0.0
djangorestframework==3.15.2
daphne==4.0.0